Versão de entrega do projeto. Com as mecânica de troca, visualização dos pokémon adversários, ataques e troca de turnos.
