+++ README.md
@@ -9,7 +9,7 @@ Dois jogadores se enfrentarão em uma batalha PvP em um jogo inspirado na franqu
 
 Os players terão um time pré definido com 6 Pokémon, dos tipos Luta, Fogo, Água, Fantasma, Psíquico e Normal, e lutarão em turnos até que um dos jogadores reduza as vidas de todos os Pokémon do adversário a 0 ou menos.
 
-Os players terão um Pokémon ativo que poderá atacar e/ou ser atacado. Todos os Pokémon poderão utilizar 6 ataques dos tipos citados acima para causar dano ao Pokémon ativo do oponente, cada Pokémon inicia a batalha com 5 vidas e sofrem dano maior ou menor de acordo com as relações de fraqueza e resistência entre os tipos. Um ataque neutro causa 2 vidas de dano, se o Pokémon utilizar um ataque do mesmo tipo que o seu a mecânica STAB (Same Type Attack Bonus) é ativada e o dano recebe um multiplicador x1.5, além disso, o ataque pode ser super efetivo ou menos efetivo, recebendo multiplicador x2 e x1/2 respectivamente. Atacar encerra o turno do jogador.
+Os players terão um Pokémon ativo que poderá atacar e/ou ser atacado. Todos os Pokémon poderão utilizar 6 ataques dos tipos citados acima para causar dano ao Pokémon ativo do oponente, cada Pokémon inicia a batalha com 5 vidas e sofrem dano maior ou menor de acordo com as relações de fraqueza e resistência entre os tipos. Um ataque neutro causa 2 vidas de dano, se o Pokémon utilizar um ataque do mesmo tipo que o seu a mecânica STAB (Same Type Attack Bonus) é ativada e o dano recebe um multiplicador x1.5, além disso, o ataque pode ser super efetivo, menos efetivo ou não efetivo, recebendo multiplicador x2, x1/2 e x0 respectivamente. Atacar encerra o turno do jogador.
 
 Alguns ataques precisarão do acionamento de um sensor específico com limite de tempo após a seleção para concretizar o ataque e acertar o Pokémon adversário, do contrário o ataque será considerado errado e não causará danos. Ataques do tipo Normal não precisarão de sensores e causarão dano automaticamente de acordo com as mecânicas de STAB e Fraqueza e Resistência. Os tipos de ataque a seguir precisarão dos seus respectivos sensores: Luta - Sensor de Distância Ultrassônico; Água - Sensor de Chuva; Fogo - Sensor de Temperatura; Psíquico - Sensor Infravermelho; Fantasma - Sensor de Presença.
 
@@ -24,14 +24,10 @@ Sensor de Chuva - Ataque de Água (1 pa);
 Sensor de Presença - Ataque Fantasma (1 pd);
 Sensor Infravermelho - Ataque Psíquico (1 pa);
 
-Joystick - Selecionar Ataques (2 pd);
 Botão - Selecionar e Cancelar Ações (1 pd);
 Buzzer - Efeitos Sonoros (1 pd pwm);
 Led RGB - Efeitos Visuais (3 pd pwm);
 LCD - Pixel Art e Visor (2 pa);
-
-Servo Motor - Acionar Hélice (1 pd);
-Hélice - Efeitos de Vento (não ocupa);
 
 ## Sprites Utilizados
